backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyMuPDF==1.23.26
google-generativeai==0.3.2
transformers==4.35.2
numpy==1.24.3
python-multipart==0.0.6
Werkzeug==2.3.7
requests==2.31.0
faiss-cpu==1.7.4
azure-cognitiveservices-speech==1.34.0
